Support multiple potential ssh ports
The "in production" or "initial setup" state of target hosts isn't
always known from the network-perspective of the control-host.
This can lead to some inventory-ambiguity over which port should
be utilized for ssh access to subject-hosts.

For example, a bastion host may have an alternate/non-standard ssh port
number, but only after it's fully configured and placed in production.

Support this use-case with this role, by accepting a list of possible
ssh ports for consideration.  Monitor, and select the first one that
opens up for use.  Assuming the test command is successful, optionally
allow setting ``ansible_port`` for the remainder of the play, for this
host.

Update CI tests to verify this port-searching behavior is functional.

Signed-off-by: Chris Evich <cevich@redhat.com>